+++ tests/DbEx.Test.Empty/Persistence/JsonConverter.cs
@@ -2,7 +2,7 @@
 
 namespace DbEx.Test.Empty.Persistence
 {
-    public class JsonConverter<T>() : ValueConverter<T, string?>(v => System.Text.Json.JsonSerializer.Serialize(v, (System.Text.Json.JsonSerializerOptions?)null), v => System.Text.Json.JsonSerializer.Deserialize<T>(v, (System.Text.Json.JsonSerializerOptions?)null)!)
+    public class JsonConverter<T>() : ValueConverter<T, string?>(v => System.Text.Json.JsonSerializer.Serialize(v, (System.Text.Json.JsonSerializerOptions?)null), v => System.Text.Json.JsonSerializer.Deserialize<T>(v!, (System.Text.Json.JsonSerializerOptions?)null)!)
     {
     }
 }